add PATCH /song/main/{songId}

+++ src/main/java/com/example/lpiloguebe/controller/DiaryController.java
@@ -28,7 +28,11 @@ public class DiaryController {
 
     @DeleteMapping("/{diaryId}")
     public ResponseEntity<?> deleteDiary(@PathVariable Long diaryId) {
+        if(diaryId == null) {
+            return new ResponseEntity<>("입력한 파라미터가 없습니다.", HttpStatus.BAD_REQUEST);
+        }
         diaryService.deleteDiary(diaryId);
         return new ResponseEntity<>("일기 삭제 완료", HttpStatus.OK);
     }
+
 }
